+++ core/src/main/scala/com/krux/hyperion/expression/Parameter.scala
@@ -1,101 +1,74 @@
 package com.krux.hyperion.expression
 
 import scala.language.implicitConversions
-import scala.reflect.runtime.universe._
 
-import org.joda.time.{DateTime, DateTimeZone}
+import org.joda.time.DateTime
 
-import com.krux.hyperion.adt.{HString, HDouble, HInt, HS3Uri, HDuration, HDateTime, HBoolean}
+import com.krux.hyperion.adt.{ HString, HDouble, HInt, HS3Uri, HDuration, HDateTime, HBoolean }
 import com.krux.hyperion.aws.AdpParameter
 import com.krux.hyperion.common.S3Uri
-import com.krux.hyperion.DataPipelineDef
 
 /**
- * @note need to add support for isOptional, allowedValues and isArray
+ * Defines and builds Parameter and returns the specific type instead of the paraent type.
  */
-case class Parameter[T : TypeTag] private (
-  id: String,
-  description: Option[String],
-  isEncrypted: Boolean
-)(implicit pv: ParameterValues) extends Evaluatable[T] { self =>
+trait ParameterBuilder[T, +Self <: Parameter[T] with ParameterBuilder[T, Self]] { self: Self =>
 
-  final val name = if (isEncrypted) s"*my_$id" else s"my_$id"
+  def updateParameterFields(fields: ParameterFields): Self
 
-  def value: Option[T] = pv.getValue(this)
-
-  def withValue(newValue: T): Parameter[T] = {
-    pv.setValue(this, newValue)
-    this
+  def withValue(newValue: T): Self = {
+    parameterFields.pv.setValue(self, newValue)
+    self
   }
 
-  def withValueFromString(newValue: String): Parameter[T] = withValue {
-    val v = typeOf[T] match {
-      case t if t <:< typeOf[Int] => newValue.toInt
-      case t if t <:< typeOf[Double] => newValue.toDouble
-      case t if t <:< typeOf[String] => newValue
-      case t if t <:< typeOf[Boolean] => newValue.toBoolean
-      case t if t <:< typeOf[DateTime] => new DateTime(newValue, DateTimeZone.UTC)
-      case t if t <:< typeOf[Duration] => Duration(newValue)
-      case t if t <:< typeOf[S3Uri] => S3Uri(newValue)
-      case _ => throw new RuntimeException("Unsupported parameter type")
-    }
-    v.asInstanceOf[T]
+  /**
+   * This needs to be implemented as a function (instead of a method) as we need to store it with
+   * the class at the time the implicit parseString for the type is availble (when the parameter
+   * instance is created). Since the calling of this function is mainly used when type is not
+   * available such as in {{{List[Parameter[_]]}}}.
+   */
+  def parseString: (String) => T
+
+  def withValueFromString(stringValue: String) = withValue(parseString(stringValue))
+
+  def withDescription(desc: String): Parameter[T] = updateParameterFields {
+    implicit val pv = parameterFields.pv
+    parameterFields.copy(description = Option(desc))
   }
 
-  def withDescription(desc: String): Parameter[T] = copy(description = Option(desc))
-  def encrypted: Parameter[T] = copy(isEncrypted = true)
+}
+
+/**
+ * The Parameter class which hides the ParameterBuilder class. Note that the parameter abstract
+ * class also belongs to the GenericParameter type class
+ */
+sealed abstract class Parameter[T : GenericParameter] extends ParameterBuilder[T, Parameter[T]] {
+
+  val env = implicitly[GenericParameter[T]]
+
+  def parameterFields: ParameterFields
+
+  def isEncrypted: Boolean
+
+  def id = parameterFields.id
+
+  def description = parameterFields.description
+
+  def name = if (isEncrypted) s"*my_$id" else s"my_$id"
+
+  def value: Option[T] = parameterFields.pv.getValue(this)
 
   def isEmpty: Boolean = value.isEmpty
 
   def evaluate(): T = value.get
 
-  def ref: TypedExpression = typeOf[T] match {
-    case t if t <:< typeOf[Int] => new IntExp with Evaluatable[Int] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[Int]
-    }
-    case t if t <:< typeOf[Double] => new DoubleExp with Evaluatable[Double] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[Double]
-    }
-    case t if t <:< typeOf[String] => new StringExp with Evaluatable[String] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[String]
-    }
-    case t if t <:< typeOf[Boolean] => new BooleanExp with Evaluatable[Boolean] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[Boolean]
-    }
-    case t if t <:< typeOf[DateTime] => new DateTimeExp with Evaluatable[DateTime] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[DateTime]
-    }
-    case t if t <:< typeOf[Duration] => new DurationExp with Evaluatable[Duration] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[Duration]
-    }
-    case t if t <:< typeOf[S3Uri] => new S3UriExp with Evaluatable[HS3Uri] {
-      def content = name
-      def evaluate() = self.evaluate().asInstanceOf[S3Uri]
-    }
-    case _ => throw new RuntimeException("Unsupported parameter type")
-  }
+  def ref: env.Exp = env.ref(this)
 
-  def `type`: String = typeOf[T] match {
-    case t if t <:< typeOf[Int] => "Integer"
-    case t if t <:< typeOf[Double] => "Double"
-    case t if t <:< typeOf[String] => "String"
-    case t if t <:< typeOf[Boolean] => "String"
-    case t if t <:< typeOf[DateTime] => "String"
-    case t if t <:< typeOf[Duration] => "String"
-    case t if t <:< typeOf[S3Uri] => "AWS::S3::ObjectKey"
-    case _ => throw new RuntimeException("Unsupported parameter type")
-  }
+  def `type`: ParameterType.Value = env.`type`
 
   def serialize: Option[AdpParameter] = Option(
     AdpParameter(
       id = name,
-      `type` = `type`,
+      `type` = `type`.toString,
       description = description,
       optional = HBoolean.False.serialize,
       allowedValues = None,
@@ -110,9 +83,17 @@ case class Parameter[T : TypeTag] private (
 
 object Parameter {
 
-  def apply[T : TypeTag](id: String)(implicit pv: ParameterValues): Parameter[T] = new Parameter[T](id, None, false)
+  def apply[T : GenericParameter](id: String)(implicit pv: ParameterValues): UnencryptedParameter[T] =
+    new UnencryptedParameter(ParameterFields(id = id))
 
-  def apply[T : TypeTag](id: String, value: T)(implicit pv: ParameterValues) = new Parameter[T](id, None, false).withValue(value)
+  def apply[T : GenericParameter](id: String, value: T)(implicit pv: ParameterValues): UnencryptedParameter[T] =
+    new UnencryptedParameter(ParameterFields(id = id)).withValue(value)
+
+  def unencrypted[T : GenericParameter](id: String)(implicit pv: ParameterValues): UnencryptedParameter[T] =
+    apply(id)
+
+  def encrypted[T : GenericParameter](id: String)(implicit pv: ParameterValues): EncryptedParameter[T] =
+    new EncryptedParameter(ParameterFields(id = id))
 
   implicit def stringParameter2HString(p: Parameter[String]): HString = HString(
     Right(
@@ -167,4 +148,33 @@ object Parameter {
       }
     )
   )
+
+}
+
+/**
+ * UnencryptedParameter is subtype of Parameter and belongs to the type class GenericParameter
+ */
+case class UnencryptedParameter[T : GenericParameter](parameterFields: ParameterFields)
+  extends Parameter[T]
+  with ParameterBuilder[T, UnencryptedParameter[T]] {
+
+  val parseString = implicitly[GenericParameter[T]].parseString
+
+  def updateParameterFields(fields: ParameterFields) = copy(parameterFields = fields)
+  val isEncrypted = false
+
+  def encrypted: EncryptedParameter[T] = new EncryptedParameter[T](parameterFields)
+}
+
+/**
+ * EncryptedParameter is subtype of Parameter and belongs to the type class GenericParameter
+ */
+case class EncryptedParameter[T : GenericParameter](parameterFields: ParameterFields)
+  extends Parameter[T]
+  with ParameterBuilder[T, EncryptedParameter[T]] {
+
+  val parseString = implicitly[GenericParameter[T]].parseString
+
+  def updateParameterFields(fields: ParameterFields) = copy(parameterFields = fields)
+  val isEncrypted = true
 }