+++ core/src/main/scala/com/krux/hyperion/expression/ParameterFields.scala
@@ -1,0 +1,6 @@
+package com.krux.hyperion.expression
+
+case class ParameterFields(
+  id: String,
+  description: Option[String] = None
+)(implicit val pv: ParameterValues)