+++ core/src/main/scala/com/krux/hyperion/expression/ParameterValues.scala
@@ -1,6 +1,6 @@
 package com.krux.hyperion.expression
 
-import scala.collection.mutable.{Map => MutableMap}
+import scala.collection.mutable.{ Map => MutableMap }
 
 class ParameterValues private (valueMap: MutableMap[String, Any]) {